hw-10 исправила ошибку
