+++ src/pages/PagePost/index.ts
@@ -1,2 +1,1 @@
-export { default } from './PagePost';
-export type { PageType } from './types';
+export { default } from './Post';
